Dashboard

+++ pizzashop.repository/Models/Ordertaxmapping.cs
@@ -11,7 +11,7 @@ public partial class Ordertaxmapping
 
     public int Taxid { get; set; }
 
-    public int? Taxvalue { get; set; }
+    public decimal? Taxvalue { get; set; }
 
     public virtual Order Order { get; set; } = null!;
